Feature/social

+++ app/src/main/java/com/toyou/toyouandroid/data/create/dto/response/AnswerPost.kt
@@ -3,5 +3,5 @@ package com.toyou.toyouandroid.data.create.dto.response
 import com.google.gson.annotations.SerializedName
 
 data class AnswerPost(
-    @SerializedName("cardId") val id : Long
+    @SerializedName("cardId") val id : Int
 )
